+++ src/Louis/Threading/ValueTaskUtility.cs
@@ -27,7 +27,7 @@ public static class ValueTaskUtility
     /// <exception cref="ArgumentNullException"><paramref name="valueTasks"/> is <see langword="null"/>.</exception>
     public static async ValueTask WhenAll(IEnumerable<ValueTask> valueTasks)
     {
-        _ = Require.NotNull(valueTasks);
+        _ = Validated.NotNull(valueTasks);
         var pendingTasks = valueTasks.Where(vt => !vt.IsCompletedSuccessfully).Select(vt => vt.AsTask()).ToList();
         if (pendingTasks.Count > 0)
         {